+++ src/main/java/app/services/FeedbackService.java
@@ -2,6 +2,7 @@ package app.services;
 
 import app.InvalidDataException;
 import app.UserContext;
+import app.entities.enums.Roles;
 import app.generated.jooq.tables.pojos.Feedback;
 import app.repository.FeedbackRepository;
 import lombok.RequiredArgsConstructor;
@@ -27,18 +28,31 @@ public class FeedbackService {
     }
 
     public void createFeedback(Feedback feedback) {
+        feedback.setAccountId(userContext.getAccountId());
+
         validateCreateFeedback(feedback);
         feedback.setFeedbackId(repository.getNextFeedbackId());
         repository.createFeedback(feedback);
     }
 
     public void deleteFeedback(Long feedbackId) {
+        if (Roles.ROLE_USER.getId().equals(userContext.getRoleId().intValue())) {
+            Feedback feedbackToDelete = repository.getById(feedbackId);
+            if (!feedbackToDelete.getAccountId().equals(userContext.getAccountId())) {
+                throw new InvalidDataException("Feedback is not accessible by this account");
+            }
+        }
         repository.deleteFeedback(feedbackId);
     }
 
     public void editFeedback(Feedback feedback) {
+        feedback.setAccountId(userContext.getAccountId());
+
         validateCreateFeedback(feedback);
         Feedback oldFeedback = repository.getById(feedback.getFeedbackId());
+        if (Roles.ROLE_USER.getId().equals(userContext.getRoleId().intValue()) && !oldFeedback.getAccountId().equals(userContext.getAccountId())) {
+            throw new InvalidDataException("Feedback is not accessible by this account");
+        }
         deleteFeedback(oldFeedback.getFeedbackId());
         repository.createFeedback(feedback);
     }
@@ -47,12 +61,6 @@ public class FeedbackService {
         Map<String, String> errors = new HashMap<>();
         if (StringUtils.isBlank(feedback.getComment())) {
             errors.put("comment", "Comment can't be null");
-        }
-
-        if (feedback.getAccountId() == null) {
-            errors.put("account", "Account can't be null");
-        } else if (!feedback.getAccountId().equals(userContext.getAccountId())) {
-            errors.put("account", "Account id mismatch");
         }
 
         if (!errors.isEmpty()) {
